added notify_completed_jobs

+++ sources/builtins/ft_fg.c
@@ -6,7 +6,7 @@
 /*   By: mbarutel <mbarutel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/01 17:09:07 by mbarutel          #+#    #+#             */
-/*   Updated: 2023/02/03 16:01:38 by mbarutel         ###   ########.fr       */
+/*   Updated: 2023/02/03 16:57:09 by mbarutel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -50,6 +50,11 @@ int	ft_fg(t_shell *sh, char **cmd)
 		ft_putstr("fg: no such job\n");
 	else
 	{
+		if (job->status == DONE || job->status == TERMINATED)
+		{
+			display_job_node(sh, job);
+			return (0);
+		}	
 		display_pipeline_cmd(job);
 		if (job->status == STOPPED || job->status == SUSPENDED)
 			killpg(job->gpid, SIGCONT);
@@ -57,8 +62,6 @@ int	ft_fg(t_shell *sh, char **cmd)
 				exit (1); // this needs to be proper exit
 		transfer_to_fg(sh, job);
 		waitpid(job->gpid, &status, WUNTRACED);
-		if (job->status == DONE)
-			bg_node_delete(sh, &job);
 	}
 	return (0);
 }
